Moves structure score tests to a single file; Adds test to Hillclimb to test all the structure score methods
