+++ pgmpy/tests/test_estimators/test_StructureScore.py
@@ -1,0 +1,139 @@
+import unittest
+
+import pandas as pd
+
+from pgmpy.models import BayesianModel
+from pgmpy.estimators import BDeuScore, BDsScore, BicScore, K2Score
+
+
+class TestBDeuScore(unittest.TestCase):
+    def setUp(self):
+        self.d1 = pd.DataFrame(
+            data={"A": [0, 0, 1], "B": [0, 1, 0], "C": [1, 1, 0], "D": ["X", "Y", "Z"]}
+        )
+        self.m1 = BayesianModel([("A", "C"), ("B", "C"), ("D", "B")])
+        self.m2 = BayesianModel([("C", "A"), ("C", "B"), ("A", "D")])
+
+        # data_link - "https://www.kaggle.com/c/titanic/download/train.csv"
+        self.titanic_data = pd.read_csv(
+            "pgmpy/tests/test_estimators/testdata/titanic_train.csv"
+        )
+        self.titanic_data2 = self.titanic_data[["Survived", "Sex", "Pclass"]]
+
+    def test_score(self):
+        self.assertAlmostEqual(BDeuScore(self.d1).score(self.m1), -9.907103407446435)
+        self.assertEqual(BDeuScore(self.d1).score(BayesianModel()), 0)
+
+    def test_score_titanic(self):
+        scorer = BDeuScore(self.titanic_data2, equivalent_sample_size=25)
+        titanic = BayesianModel([("Sex", "Survived"), ("Pclass", "Survived")])
+        self.assertAlmostEqual(scorer.score(titanic), -1892.7383393910427)
+        titanic2 = BayesianModel([("Pclass", "Sex")])
+        titanic2.add_nodes_from(["Sex", "Survived", "Pclass"])
+        self.assertLess(scorer.score(titanic2), scorer.score(titanic))
+
+    def tearDown(self):
+        del self.d1
+        del self.m1
+        del self.m2
+        del self.titanic_data
+        del self.titanic_data2
+
+
+class TestBDsScore(unittest.TestCase):
+    def setUp(self):
+        """Example taken from https://arxiv.org/pdf/1708.00689.pdf"""
+        self.d1 = pd.DataFrame(
+            data={
+                "X": [0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0],
+                "Y": [0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1],
+                "Z": [0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1],
+                "W": [0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1],
+            }
+        )
+        self.m1 = BayesianModel([("W", "X"), ("Z", "X")])
+        self.m1.add_node("Y")
+        self.m2 = BayesianModel([("W", "X"), ("Z", "X"), ("Y", "X")])
+
+    def test_score(self):
+        self.assertAlmostEqual(
+            BDsScore(self.d1, equivalent_sample_size=1).score(self.m1),
+            -36.82311976667139,
+        )
+        self.assertEqual(
+            BDsScore(self.d1, equivalent_sample_size=1).score(self.m2),
+            -45.788991276221964,
+        )
+
+    def tearDown(self):
+        del self.d1
+        del self.m1
+        del self.m2
+
+
+class TestBicScore(unittest.TestCase):
+    def setUp(self):
+        self.d1 = pd.DataFrame(
+            data={"A": [0, 0, 1], "B": [0, 1, 0], "C": [1, 1, 0], "D": ["X", "Y", "Z"]}
+        )
+        self.m1 = BayesianModel([("A", "C"), ("B", "C"), ("D", "B")])
+        self.m2 = BayesianModel([("C", "A"), ("C", "B"), ("A", "D")])
+
+        # data_link - "https://www.kaggle.com/c/titanic/download/train.csv"
+        self.titanic_data = pd.read_csv(
+            "pgmpy/tests/test_estimators/testdata/titanic_train.csv"
+        )
+        self.titanic_data2 = self.titanic_data[["Survived", "Sex", "Pclass"]]
+
+    def test_score(self):
+        self.assertAlmostEqual(BicScore(self.d1).score(self.m1), -10.698440814229318)
+        self.assertEqual(BicScore(self.d1).score(BayesianModel()), 0)
+
+    def test_score_titanic(self):
+        scorer = BicScore(self.titanic_data2)
+        titanic = BayesianModel([("Sex", "Survived"), ("Pclass", "Survived")])
+        self.assertAlmostEqual(scorer.score(titanic), -1896.7250012840179)
+        titanic2 = BayesianModel([("Pclass", "Sex")])
+        titanic2.add_nodes_from(["Sex", "Survived", "Pclass"])
+        self.assertLess(scorer.score(titanic2), scorer.score(titanic))
+
+    def tearDown(self):
+        del self.d1
+        del self.m1
+        del self.m2
+        del self.titanic_data
+        del self.titanic_data2
+
+
+class TestK2Score(unittest.TestCase):
+    def setUp(self):
+        self.d1 = pd.DataFrame(
+            data={"A": [0, 0, 1], "B": [0, 1, 0], "C": [1, 1, 0], "D": ["X", "Y", "Z"]}
+        )
+        self.m1 = BayesianModel([("A", "C"), ("B", "C"), ("D", "B")])
+        self.m2 = BayesianModel([("C", "A"), ("C", "B"), ("A", "D")])
+
+        # data_link - "https://www.kaggle.com/c/titanic/download/train.csv"
+        self.titanic_data = pd.read_csv(
+            "pgmpy/tests/test_estimators/testdata/titanic_train.csv"
+        )
+        self.titanic_data2 = self.titanic_data[["Survived", "Sex", "Pclass"]]
+
+    def test_score(self):
+        self.assertAlmostEqual(K2Score(self.d1).score(self.m1), -10.73813429536977)
+        self.assertEqual(K2Score(self.d1).score(BayesianModel()), 0)
+
+    def test_score_titanic(self):
+        scorer = K2Score(self.titanic_data2)
+        titanic = BayesianModel([("Sex", "Survived"), ("Pclass", "Survived")])
+        self.assertAlmostEqual(scorer.score(titanic), -1891.0630673606006)
+        titanic2 = BayesianModel([("Pclass", "Sex")])
+        titanic2.add_nodes_from(["Sex", "Survived", "Pclass"])
+        self.assertLess(scorer.score(titanic2), scorer.score(titanic))
+
+    def tearDown(self):
+        del self.d1
+        del self.m1
+        del self.m2
+        del self.titanic_data
+        del self.titanic_data2
